separate options and data stores
