+++ src/components/clock.vue
@@ -1,12 +1,14 @@
 <script setup lang="ts">
 import { onMounted } from 'vue'
 import { useOptionsStore } from '../store/options'
+import { useDataStore } from '../store/data'
 import Clockface from './clockface.vue'
 
 const optionsStore = useOptionsStore()
+const dataStore = useDataStore()
 
 onMounted(() => {
-	optionsStore.getTime()
+	dataStore.getTime()
 })
 </script>
 
@@ -14,16 +16,16 @@ onMounted(() => {
 	<div class="wrapper">
 		<Clockface
 			:layout="optionsStore.time.layout"
-			:hour="optionsStore.formattedDate.hour"
-			:minute="optionsStore.formattedDate.minute"
-			:time="optionsStore.data.date"
+			:hour="dataStore.formattedDate.hour"
+			:minute="dataStore.formattedDate.minute"
+			:time="dataStore.data.date"
 		/>
 		<div class="space-small" />
 		<div class="context-label" v-if="!optionsStore.time.use24Hour">
-			{{ optionsStore.formattedDate.descriptor }}
+			{{ dataStore.formattedDate.descriptor }}
 		</div>
 		<div v-else class="context-label">today is</div>
-		<div class="date">{{ optionsStore.formattedDate.today }}</div>
+		<div class="date">{{ dataStore.formattedDate.today }}</div>
 	</div>
 </template>
