+++ src/utils/theme.ts
@@ -1,0 +1,8 @@
+export function applyTheme(theme: string) {
+	document.documentElement.className = theme
+}
+
+export function setTheme(theme: string) {
+	applyTheme(theme)
+	localStorage.setItem('options/theme/color', theme)
+}
