+++ src/utils/weather-conditions.ts
